Drobné úpravy

+++ prepare.py
@@ -25,6 +25,7 @@ def prepareUserAreas():
 
 def prepare():
 	prepareUserAreas()
+	
 	from makerfuncs import config, download as d
 	import update as u
 
@@ -59,20 +60,25 @@ def prepare():
 		try:
 			os.mkdir(data[item])
 		except:
+			print("Directory " + data[item] + " already exists")
 			pass
 
-	# TODO phyghtmap
 
+	if (not os.path.isdir(data['sea'])):
+		d.download('http://osm.thkukuk.de/data/sea-latest.zip', './sea.zip')
+		with zipfile.ZipFile('./sea.zip', 'r') as zipRef:
+			zipRef.extractall(data['sea'])
+		os.remove('./sea.zip')
+	else:
+		print("Directoty " + data['sea'] + "already exists - skipping...")
 
-	d.download('http://osm.thkukuk.de/data/sea-latest.zip', './sea.zip')
-	with zipfile.ZipFile('./sea.zip', 'r') as zipRef:
-		zipRef.extractall(data['sea'])
-	os.remove('./sea.zip')
-
-	d.download('http://osm.thkukuk.de/data/bounds-latest.zip', './bounds.zip')
-	with zipfile.ZipFile('./bounds.zip', 'r') as zipRef:
-		zipRef.extractall(data['bounds'])
-	os.remove('./bounds.zip')
+	if (not os.path.isdir(data['bounds'])):
+		d.download('http://osm.thkukuk.de/data/bounds-latest.zip', './bounds.zip')
+		with zipfile.ZipFile('./bounds.zip', 'r') as zipRef:
+			zipRef.extractall(data['bounds'])
+		os.remove('./bounds.zip')
+	else:
+		print("Directoty " + data['bounds'] + "already exists - skipping...")
 
 
 	data['splitter'] = 0
